Moved framework elements into src
