+++ apps/e2e/package.json
@@ -32,5 +32,10 @@
     "ts-node": "10.9.2",
     "typescript": "5.8.3"
   },
-  "dependencies": {}
+  "dependencies": {
+    "@tryghost/debug": "^0.1.34",
+    "@tryghost/errors": "^1.3.7",
+    "@tryghost/logging": "^2.4.21",
+    "mailhog": "^4.16.0"
+  }
 }
