+++ apps/e2e/src/types/logging.d.ts
@@ -1,0 +1,1 @@
+declare module '@tryghost/logging';
