+++ apps/e2e/test/e2e/admin.spec.ts
@@ -1,15 +1,14 @@
-import { test, expect } from '../test-fixtures';
+import {test} from '../../src/test-fixtures';
 
 test.describe('Admin Login', () => {
-  test('should allow login with environment variable credentials and 2FA',
-    async ({ loginPage, twoFactorAuthPage, dashboardPage }) => {
+    test('should allow login with environment variable credentials and 2FA',
+        async ({loginPage, twoFactorAuthPage, dashboardPage}) => {
+            await loginPage.goto();
 
-    await loginPage.goto();
+            await loginPage.login();
 
-    await loginPage.login();
+            await twoFactorAuthPage.complete2FA();
 
-    await twoFactorAuthPage.complete2FA();
-
-    await dashboardPage.expectCurrentUrl();
-  });
+            await dashboardPage.expectCurrentUrl();
+        });
 });
