+++ apps/e2e/test/e2e/frontend.spec.ts
@@ -1,6 +1,6 @@
-import { test, expect } from '../test-fixtures'; // Use our custom test and expect
+import {expect, test} from '../../src/test-fixtures';
 
-test('has correct title', async ({ page, appUrls }) => {
-  await page.goto(appUrls.baseURL);
-  await expect(page).toHaveTitle('[DEV]');
+test('has correct title', async ({page, appUrls}) => {
+    await page.goto(appUrls.baseURL);
+    await expect(page).toHaveTitle('[DEV]');
 });
